+++ cmd/golivepilot/main.go
@@ -22,7 +22,7 @@ import (
 )
 
 var (
-	version = "0.1.58"
+	version = "0.1.59"
 )
 
 func main() {